+++ apps/storybook-host/src/rhf-core/form-container.stories.tsx
@@ -1,4 +1,5 @@
-import { FormContainer, HtmlButtonElement, HtmlInputElement } from '@piplup/rhf-core';
+import { FormContainer } from '@piplup/rhf-core';
+import { HtmlButtonElement, HtmlInputElement } from '@piplup/rhf-core/html';
 import { action } from '@storybook/addon-actions';
 import { type Meta, type StoryObj } from '@storybook/react';
