+++ apps/storybook-host/src/rhf-core/html/form-helper-text.stories.tsx
@@ -1,9 +1,9 @@
+import { FormContainer } from '@piplup/rhf-core';
 import {
   HtmlInputElement,
   HtmlButtonElement,
-  FormContainer,
   HtmlFormHelperTextElement,
-} from '@piplup/rhf-core';
+} from '@piplup/rhf-core/html';
 import { action } from '@storybook/addon-actions';
 import { type Meta, type StoryObj } from '@storybook/react';
 
@@ -13,7 +13,7 @@ import { type Meta, type StoryObj } from '@storybook/react';
  * ## <span className="docs-h2">Import</span>
  *
  * ```tsx
- * import { HtmlFormHelperTextElement } from '@piplup/rhf-core';
+ * import { HtmlFormHelperTextElement } from '@piplup/rhf-core/html';
  * ```
  */
 const meta: Meta<typeof HtmlFormHelperTextElement> = {
