+++ apps/storybook-host/src/rhf-core/html/form-label.stories.tsx
@@ -1,10 +1,10 @@
+import { FormContainer } from '@piplup/rhf-core';
 import {
   HtmlInputElement,
   HtmlButtonElement,
-  FormContainer,
   HtmlFormHelperTextElement,
   HtmlFormLabelElement,
-} from '@piplup/rhf-core';
+} from '@piplup/rhf-core/html';
 import { action } from '@storybook/addon-actions';
 import { type Meta, type StoryObj } from '@storybook/react';
 
@@ -14,7 +14,7 @@ import { type Meta, type StoryObj } from '@storybook/react';
  * ## <span className="docs-h2">Import</span>
  *
  * ```tsx
- * import { HtmlFormLabelElement } from '@piplup/rhf-core';
+ * import { HtmlFormLabelElement } from '@piplup/rhf-core/html';
  * ```
  */
 const meta: Meta<typeof HtmlFormHelperTextElement> = {
