+++ apps/storybook-host/src/rhf-core/html/input.stories.tsx
@@ -1,9 +1,9 @@
+import { FormContainer } from '@piplup/rhf-core';
 import {
-  FormContainer,
   HtmlButtonElement,
   HtmlFormHelperTextElement,
   HtmlInputElement,
-} from '@piplup/rhf-core';
+} from '@piplup/rhf-core/html';
 import { action } from '@storybook/addon-actions';
 import { type Meta, type StoryObj } from '@storybook/react';
 
@@ -13,7 +13,7 @@ import { type Meta, type StoryObj } from '@storybook/react';
  * ## <span className="docs-h2">Import</span>
  *
  * ```tsx
- * import { HtmlInputElement } from '@piplup/rhf-core';
+ * import { HtmlInputElement } from '@piplup/rhf-core/html';
  * ```
  */
 const meta: Meta<typeof HtmlInputElement> = {
