+++ apps/storybook-host/src/rhf-core/html/textarea.stories.tsx
@@ -1,10 +1,9 @@
+import { FormContainer, type FormContainerProps } from '@piplup/rhf-core';
 import {
-  FormContainer,
-  type FormContainerProps,
   HtmlButtonElement,
   HtmlFormHelperTextElement,
   HtmlTextareaElement,
-} from '@piplup/rhf-core';
+} from '@piplup/rhf-core/html';
 import { action } from '@storybook/addon-actions';
 import { type Meta, type StoryObj } from '@storybook/react';
 
@@ -14,7 +13,7 @@ import { type Meta, type StoryObj } from '@storybook/react';
  * ## <span className="docs-h2">Import</span>
  *
  * ```tsx
- * import { HtmlTextareaElement } from '@piplup/rhf-core';
+ * import { HtmlTextareaElement } from '@piplup/rhf-core/html';
  * ```
  */
 const meta: Meta<typeof HtmlTextareaElement> = {
