+++ packages/rhf-adapters/src/mui/lab/loading-button/adapter.ts
@@ -1,5 +1,5 @@
 import type * as React from 'react';
-import { useHtmlButtonAdapter, type UseHtmlButtonAdapterProps } from '@piplup/rhf-core';
+import { useHtmlButtonAdapter, type UseHtmlButtonAdapterProps } from '@piplup/rhf-core/html';
 import { useFormState, type FieldValues } from 'react-hook-form';
 
 export interface UseMuiLoadingButtonAdapterProps<TFieldValues extends FieldValues = FieldValues>
