+++ packages/rhf-adapters/src/mui/material/fab/adapter.ts
@@ -1,5 +1,5 @@
 import type * as React from 'react';
-import { useHtmlButtonAdapter, type UseHtmlButtonAdapterProps } from '@piplup/rhf-core';
+import { useHtmlButtonAdapter, type UseHtmlButtonAdapterProps } from '@piplup/rhf-core/html';
 import { type FieldValues } from 'react-hook-form';
 
 export type UseMuiFabAdapterProps<TFieldValues extends FieldValues = FieldValues> =
