+++ packages/rhf-adapters/src/mui/material/input-label/adapter.ts
@@ -1,4 +1,4 @@
-import { useHtmlFormLabelAdapter, type UseHtmlFormLabelProps } from '@piplup/rhf-core';
+import { useHtmlFormLabelAdapter, type UseHtmlFormLabelProps } from '@piplup/rhf-core/html';
 import { type FieldPath, type FieldValues } from 'react-hook-form';
 
 export interface UseMuiInputLabelAdapterProps<
