+++ packages/rhf-core/src/index.ts
@@ -21,39 +21,3 @@ export {
   type UseFormResetResult,
 } from './hooks/use-form-reset';
 export { useFormStateAdapter, type UseFormStateAdapterProps } from './hooks/use-form-state-adapter';
-
-export {
-  HtmlButtonClasses,
-  HtmlButtonElement,
-  type HtmlButtonElementProps,
-  useHtmlButtonAdapter,
-  type UseHtmlButtonAdapterProps,
-} from './html/button';
-export {
-  HtmlFormHelperTextClasses,
-  HtmlFormHelperTextElement,
-  type HtmlFormHelperTextElementProps,
-  useHtmlFormHelperTextAdapter,
-  type UseHtmlFormHelperTextProps,
-} from './html/form-helper-text';
-export {
-  HtmlFormLabelClasses,
-  HtmlFormLabelElement,
-  type HtmlFormLabelElementProps,
-  useHtmlFormLabelAdapter,
-  type UseHtmlFormLabelProps,
-} from './html/form-label';
-export {
-  HtmlInputClasses,
-  HtmlInputElement,
-  type HtmlInputElementProps,
-  useHtmlInputAdapter,
-  type UseHtmlInputAdapterProps,
-} from './html/input';
-export {
-  HtmlTextareaClasses,
-  HtmlTextareaElement,
-  type HtmlTextareaElementProps,
-  useHtmlTextareaAdapter,
-  type UseHtmlTextareaAdapterProps,
-} from './html/textarea';
