+++ tsconfig.base.json
@@ -23,6 +23,7 @@
         "packages/rhf-adapters/src/mui/material/mui-material.ts"
       ],
       "@piplup/rhf-core": ["packages/rhf-core/src/index.ts"],
+      "@piplup/rhf-core/html": ["packages/rhf-core/src/html/html.ts"],
       "@piplup/storybook-host": ["apps/storybook-host/src/index.ts"],
       "@piplup/use-event-listener": ["packages/use-event-listener/src/index.ts"],
       "@piplup/use-local-storage": ["packages/use-local-storage/src/index.ts"],
